svg/text/parsing/text-decoration-fill-invalid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>SVG Text: parsing text-decoration-fill with invalid values</title><metadata><h:link href="https://svgwg.org/svg2-draft/text.html#TextDecorationFillProperty" rel="help"/><h:meta content="text-decoration-fill supports only the paint grammar 'none | color | url [none | color]? | context-fill | context-stroke'." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_invalid_value("text-decoration-fill", "auto");
test_invalid_value("text-decoration-fill", "none red");
test_invalid_value("text-decoration-fill", 'none url("https://example.com/")');
test_invalid_value("text-decoration-fill", 'red url("https://example.com/")');
test_invalid_value("text-decoration-fill", 'url("https://example.com/") none red');

  ]]></script></svg>
